densenet121/requirements.txt
tensorflow==2.0.0
Pillow==7.0.0